CIF-1328: Created NavigationContext refactored AuthBar to use context
